% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_package.R
\name{analyse_package}
\alias{analyse_package}
\title{Analyse the function structure in an R package}
\usage{
analyse_package(directory, is_local = FALSE)
}
\arguments{
\item{directory}{A string indicating the directory of the package to be analysed}

\item{is_local}{A logical indicating whether or not the package is local or a URL (this decides the use of "\" or "/" respectively).}
}
\value{
A dataframe of all the .R files found in an R package
}
\description{
The \code{analyse_package} function finds and analyses all the .R files it can find in a package folder given by \code{directory}.
}
