% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Open ExploreR dashboard}
\usage{
explore(theme = "sandstone")
}
\arguments{
\item{theme}{Shiny theme from shinythemes you wish to set. Defaults to 'sandstone'.}
}
\description{
This function loads the shiny app.
}
\details{
Roxygen tag to export functions
ensures functions in this document get related to the NAMESPACE when running devtools::document()
}
