% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{missing_values}
\alias{missing_values}
\title{Missing Values}
\usage{
missing_values(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
counts the number of missing values present in a data frame
and reports back number and percent missing for each variable
}
\examples{
data <- data.frame("names" = c("Rachel", "Jim", "Milos", NA),
                   "numbers" = c(2, NA, 6, 8),
                   "truths" = c(NA, TRUE, TRUE, FALSE))
missing_values(data)
}
