\name{na.replace}
\alias{na.replace}
\title{Sostituzione dei dati mancanti}
\description{
    Da descrivere
}
\usage{
na.replace(x, margin = 2, fun = mean, ...)
}
\arguments{
    \item{x}{Da descrivere}
    \item{margin}{Da descrivere}
    \item{fun}{Da descrivere}
    \item{\dots}{Da descrivere}
}
\details{
    Da descrivere
}
\value{
    Da descrivere
}
\references{
    Da descrivere
}
\author{
Davide Massidda <davide.massidda@gmail.com>
}
\examples{
# Da descrivere
}