% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pbernd}
\alias{pbernd}
\title{Bernoulli
Binomial with N = 1. One run of an experiment with a binary outcome (e.g., toss a coin; can get either heads or tails)}
\usage{
pbernd(x = 0:1, p = 0.8)
}
\description{
Bernoulli
Binomial with N = 1. One run of an experiment with a binary outcome (e.g., toss a coin; can get either heads or tails)
}
