% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pbetad}
\alias{pbetad}
\title{Beta
Continuous distribution bounded to lie between 0 and 1. Commonly used as a distribution describing p, the per-trial success probability for Bernoulli trials.}
\usage{
pbetad(x = seq(0, 1, by = 0.01), alpha = 5, beta = 2)
}
\description{
Beta
Continuous distribution bounded to lie between 0 and 1. Commonly used as a distribution describing p, the per-trial success probability for Bernoulli trials.
}
