% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pbinomd}
\alias{pbinomd}
\title{Discrete Variables
Binomial
Number of successes in n trials}
\usage{
pbinomd(x = 1:10, n = 5, p = 0.5)
}
\description{
Discrete Variables
Binomial
Number of successes in n trials
}
