% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pexpd}
\alias{pexpd}
\title{Exponential
Describes the time between events in a Poisson point process}
\usage{
pexpd(x = seq(0, 1, by = 0.01), rate = 5)
}
\description{
Exponential
Describes the time between events in a Poisson point process
}
