% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pgammad}
\alias{pgammad}
\title{Gamma
A stochastic process with independent, non-negative increments having a gamma distribution with an identical scale parameter and a time-dependent shape parameter.}
\usage{
pgammad(x = seq(0, 1, by = 0.01), alpha = 5, beta = 2)
}
\description{
Gamma
A stochastic process with independent, non-negative increments having a gamma distribution with an identical scale parameter and a time-dependent shape parameter.
}
