% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pgeomd}
\alias{pgeomd}
\title{Geometric
In a sequence of independent and identically distributed Bernoulli trials with success probability p, if X = number of trials until the 1st success, then X ~ Geo(p).}
\usage{
pgeomd(x = 1:10, p = 0.5)
}
\description{
Geometric
In a sequence of independent and identically distributed Bernoulli trials with success probability p, if X = number of trials until the 1st success, then X ~ Geo(p).
}
