% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{phyperd}
\alias{phyperd}
\title{Hypergeometric
In a population of total size N with two types of individuals, Type 1 and Type 2 (M type 1's and N - M type 2's), X = number of Type 1's in a sample of size n drawn without replacement}
\usage{
phyperd(draws = 1:10, type1 = 5, type2 = 5, trials = 8)
}
\description{
Hypergeometric
In a population of total size N with two types of individuals, Type 1 and Type 2 (M type 1's and N - M type 2's), X = number of Type 1's in a sample of size n drawn without replacement
}
