\name{plot.cor.drop}
\alias{plot.cor.drop}
\title{Da descrivere}
\description{
    Da descrivere
}
\usage{
plot.cor.drop(object, vertical = FALSE, main = NULL, axis.lim = NULL, mar.bottom = 5.1, mar.left = 4.1, mar.top = 4.1, mar.right = 2.1)
}
\arguments{
  \item{object}{Da descrivere}
  \item{vertical}{Da descrivere}
  \item{main}{}
  \item{axis.lim}{Da descrivere}
  \item{mar.bottom}{Da descrivere}
  \item{mar.left}{Da descrivere}
  \item{mar.top}{Da descrivere}
  \item{mar.right}{Da descrivere}
}
\details{
    Da descrivere
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{
Davide Massidda <davide.massidda@gmail.com>
}
\note{
Da descrivere
}
\examples{
# Da descrivere
}