% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pnbinomd}
\alias{pnbinomd}
\title{Negative Binomial
In a sequence of independent and identically distributed Bernoulli trials with success probability p, if X = number of trials until the rth success, then X ~ NBin(r, p).}
\usage{
pnbinomd(x = 1:100, r = 50, p = 0.5)
}
\description{
Negative Binomial
In a sequence of independent and identically distributed Bernoulli trials with success probability p, if X = number of trials until the rth success, then X ~ NBin(r, p).
}
