% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pnormd}
\alias{pnormd}
\title{Normal
Among other process bases, according to the Central Limit Theorem, the averages of samples of observations of random variables independently drawn from independent distributions converge in distribution to the normal}
\usage{
pnormd(x = seq(0, 100, by = 0.01), mu = 25, sigma = 2)
}
\description{
Normal
Among other process bases, according to the Central Limit Theorem, the averages of samples of observations of random variables independently drawn from independent distributions converge in distribution to the normal
}
