% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{ppoisd}
\alias{ppoisd}
\title{Poisson
Probability of a fixed number of events in a given interval (often of time or space), given that events occur at a constant rate}
\usage{
ppoisd(x = 1:10, lambda = 5)
}
\description{
Poisson
Probability of a fixed number of events in a given interval (often of time or space), given that events occur at a constant rate
}
