\name{print.cor.matrix}
\alias{print.cor.matrix}
\title{Da descrivere}
\description{
    Da descrivere
}
\usage{
\method{print}{cor.matrix}(x,...)
}
\arguments{
    \item{x}{Da descrivere}
    \item{...}{Da descrivere}
}
\details{Da descrivere}
\value{
    Da descrivere
}
\author{
Davide Massidda <davide.massidda@gmail.com>
}
\examples{
# Da descrivere
}
