% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{punifd}
\alias{punifd}
\title{Continuous Variables
Uniform
Within some bounded interval, all outcomes have an equal probability of occurrence.}
\usage{
punifd(x = seq(0, 1, by = 0.01), a = 0.1, b = 0.9)
}
\description{
Continuous Variables
Uniform
Within some bounded interval, all outcomes have an equal probability of occurrence.
}
