% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pvonmisesd}
\alias{pvonmisesd}
\title{Von Mises
The maximum entropy distribution for circular data when the real and imaginary parts of the first circular moment are specified}
\usage{
pvonmisesd(x = seq(-pi, pi, by = 0.01), mu = 1, kappa = 0.1)
}
\description{
Von Mises
The maximum entropy distribution for circular data when the real and imaginary parts of the first circular moment are specified
}
