% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_96_clariostar}
\alias{read_96_clariostar}
\title{readclariostar}
\usage{
read_96_clariostar(file_path)
}
\arguments{
\item{file_path}{File path to the .csv file from the clariostar.}
}
\value{
a tibble with 3 columns, time, well and od600
}
\description{
Reads the plates from the clariostar plate reader. Hooray!
}
