% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Reset R}
\usage{
reset()
}
\description{
This function simply erases the the console, detaches all packages and removes all data from the global environment.  The purpose of which is to provide an easy command which can clean up the workspace.  Very useful after you spend some time experimenting.
}
\examples{
reset()
}

