% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Shape and Size Function}
\usage{
size(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
Gives the shape and size of a dataframe
}
\examples{
data <- data.frame("names" = c("Rachel", "Jim", "Milos", "Arzan"),
                   "numbers" = c(2, 4, 6, 8),
                   "truths" = c(TRUE, TRUE, TRUE, FALSE))
size(data)
}
