% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{sliderInput01}
\alias{sliderInput01}
\title{Tab Constructors for UI}
\usage{
sliderInput01(id, label)
}
\arguments{
\item{id}{The shiny variable used.}

\item{label}{The label for the slider.}
}
\description{
Shiny slider <0 to 1> input function
}
