\name{VPE}
\alias{VPE}
\docType{data}
\title{VPE: Vanishing Point Estimation}
\description{
    Il dataset VPE riporta dei dati simulati, creati basandosi su valori realmente registrati nel corso
    di un esperimento di psicofisica. Nell'esperimento, le persone sono state poste davanti a un monitor
    nel quale appariva un piccolo punto in movimento che percorreva una traiettoria lineare; dopo qualche
    secondo lo stimolo spariva: compito delle persone era di indicare con il mouse il punto di scomparsa
    (vanishing point). Nel dataset VPE sono riportati i dati di cinque ipotetici soggetti sperimentali.

    Ogni soggetto affrontava tre condizioni sperimentali: (a) una condizione nella quale la velocita' dello
    stimolo aumenta durante il tragitto; (b) una condizione nella quale la velocita' dello stimolo e' costante;
    (c) una condizione nella quale la velocita' dello stimolo diminuisce.
    
    Ogni partecipante affrontava quattro volte una medesima condizione sperimentale, e per ognuna viene
    registrato l'errore commesso.
}
\usage{data(VPE)}
\format{Un oggetto di tipo \code{matrix}.}
\keyword{datasets}