\name{plot.cor.drop}
\alias{plot.cor.drop}
\title{Da descrivere}
\description{
    Da descrivere
}
\usage{
\method{plot}{cor.drop}(x, vertical = FALSE, main = NULL, axis.lim = NULL,
    mar.bottom = 5.1, mar.left = 4.1, mar.top = 4.1, mar.right = 2.1, ...)
}
\arguments{
  \item{x}{Da descrivere}
  \item{vertical}{Da descrivere}
  \item{main}{Da descrivere}
  \item{axis.lim}{Da descrivere}
  \item{mar.bottom}{Da descrivere}
  \item{mar.left}{Da descrivere}
  \item{mar.top}{Da descrivere}
  \item{mar.right}{Da descrivere}
  \item{...}{Da descrivere}
}
\details{
    Da descrivere
}
\value{
    Da descrivere
}
\author{
Davide Massidda <davide.massidda@gmail.com>
}
\note{
Da descrivere
}
\examples{
# Da descrivere
}