\name{print.cor.drop}
\alias{print.cor.drop}
\title{Da descrivere}
\description{
    Da descrivere
}
\usage{
print.cor.drop(object)
}
\arguments{
    \item{object}{Da descrivere}
}
\details{
    Da descrivere
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{
Davide Massidda <davide.massidda@gmail.com>
}
\examples{
# Da descrivere
}