% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summary.R
\name{variable_summary}
\alias{variable_summary}
\title{Variable Summary}
\usage{
variable_summary(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
takes a data frame as input and provides the total quantity
of each type of variable present in the data frame
}
